separate ptree.c from prinfo.h and ptree.h

+++ include/linux/syscalls.h
@@ -80,6 +80,7 @@ union bpf_attr;
 #include <linux/quota.h>
 #include <linux/key.h>
 #include <trace/syscall.h>
+#include <linux/ptree.h>
 
 /*
  * __MAP - apply a macro to syscall arguments
@@ -940,6 +941,6 @@ asmlinkage long sys_pkey_alloc(unsigned long flags, unsigned long init_val);
 asmlinkage long sys_pkey_free(int pkey);
 asmlinkage long sys_statx(int dfd, const char __user *path, unsigned flags,
 			  unsigned mask, struct statx __user *buffer);
-asmlinkage long sys_ptree(struct prinfo *buf, int *nr);
+asmlinkage int sys_ptree(struct prinfo *buf, int *nr);
 
 #endif
